+++ CHANGELOG.md
@@ -2,5 +2,8 @@
 ----------
 - Added Helper component
 - Added Tooltip component
+- Added Slider component
 - Added helper prop to field and select
 - Hiding icons in read-only mode in select component
+- Removed babel-polyfill
+- Override colors and constants props for ValidStyles